jsgen: implement proper 64-bit integer support (#1694)
## Summary

For the JavaScript backend, make `int64`, `uint64` and other integer
types work like they do when using the other backends.

## Details

The changes to run-time behaviour are:
* `int64` and `uint64` are full 64-bit precision integers
* casting from unsigned to signed integers performs a proper bitcast
* casting float to int and vice versa performs a bitcast instead of
  being a no-op
* converting `bool` to numeric types yields either 1 or 0 instead of
  being a no-op
* converting from float to unsigned int performs a truncation +
  clamp-to-range instead of being a no-op
* converting from float to signed int clamps to range instead of
  truncating and bitcasting
* `uint32` multiplication returns the correct result for large operand
  values now
* left-shifting `uint16` and `uint8` doesn't yield invalid values
  anymore
* inverting the bits of `uint16` and `uint8` via `not` doesn't yield
  invalid values anymore

The main intention was to implement support for 64-bit integers for the
JavaScript backend, with some of the other fixes being required by
the 64-bit integer implementation.

Internally, `jsgen` is changed such that it uses different paths for
`int64` and `uint64` compared to the other integer types, both for
constants and when used as operands. The implementation for the various
64-bit integer operations is provided as procedure in the new `jsint64`
system module -- they're pure implementation as much as possible in
order for them to still work without change when switching to `asm.js`.

When used as an index, 64-bit integer are internally converted to a
`float64` (a JS `Number`) first, which keeps the behaviour the same as
it was previously.

Similarly, for range checks, 64-bit integers are first converted to
`float64`, meaning that range checks with values outside the
32-bit integer range still don't work. The reason for this is to not
having to duplicate the logic in `rtchecks` into `jsgen` for now
(`rtchecks` should be enabled for the JavaScript target eventually).

Finally:
* the 64-bit integer conversion to `JsBigInt` in the standard library
  are adjusted such that they keep the full precision
* internal adjustments are made to the `hashes` and `times` modules
  where the previous `cast` and `int` behaviour was relied on
* various tests related to casting, conversions, and bit operations are
  enabled for the JavaScript backend

+++ lib/pure/times.nim
@@ -897,7 +897,9 @@ proc toWinTime*(t: Time): int64 =
 proc getTime*(): Time {.tags: [TimeEffect], benign.} =
   ## Gets the current time as a `Time` with up to nanosecond resolution.
   when defined(js):
-    let millis = newDate().getTime()
+    # XXX: converting to float first and then int64 is a workaround for
+    #      `getTime` not returning a `float`
+    let millis = int64(float(newDate().getTime()))
     let seconds = convert(Milliseconds, Seconds, millis)
     let nanos = convert(Milliseconds, Nanoseconds,
       millis mod convert(Seconds, Milliseconds, 1).int)
